src/requirements.txt
azure-core
azure-identity
azure-search-documents
azure-ai-documentintelligence
azure-storage-blob
chainlit
langchain
openai
pathlib
prompty
python-docx
python-dotenv
semantic-kernel
tenacity
tiktoken